Added Light Investigation mode to AI
Enemy will investigate using a 3 point path, the 2nd point is where the
"light switch" is, and off the z-axis of the normal route. After the light
is turned back on, the guard resumes normal guard. While the light is
turned off the guard has a customizable "reduced" vision while it is
walking toward the switch to investigate.